refactor: reorganize vision classes into odometry package
- Move PhotonVisionCamera and VisionOdometry to the odometry package
- Update package declarations accordingly
